Removes console log
